requirements.txt
discord.py==1.4.1
dnspython==2.0.0
PyNaCl==1.4.0
async-timeout==3.0.1
requests==2.24.0
python-dotenv==0.14.0
APScheduler==3.6.0